Allow for a different default search provider for private tabs
Introduces setting: 'search.use-alternate-private-search-engine which defines whether or not to use a different
search provider, currently hardcoded to DuckDuckGo, when searching from a Private Tab.
Uses this setting when searching in the UrlBar when a private frame is active.
Setting can be changed from about:preferences or about:newtab (when in a Private Tab)

Adjusts styles of Private New Tab page to include new fonts and typography rules

Fix #11305